django-comments/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
